modify task to do resampling
