--- docs/api/tasks/pyhealth.tasks.ptbxl_diagnostic_classification.rst
@@ -1,11 +1,0 @@
-PTB-XL Diagnostic Classification
-================================
-
-.. currentmodule:: pyhealth.tasks.ptbxl_diagnostic_classification
-
-.. autoclass:: PTBXLDiagnosticClassification
-    :members:
-    :show-inheritance:
-    :exclude-members: __init__
-
-    .. automethod:: __call__
