+++ docs/api/tasks/pyhealth.tasks.ptbxl_resampling.rst
@@ -1,0 +1,11 @@
+PTB-XL Signal Resampling
+========================
+
+.. currentmodule:: pyhealth.tasks.ptbxl_resampling
+
+.. autoclass:: PTBXLResampling
+    :members:
+    :show-inheritance:
+    :exclude-members: __init__
+
+    .. automethod:: __call__
