+++ pyhealth/tasks/__init__.py
@@ -11,7 +11,7 @@ from .cardiology_detect import (
 )
 from .chestxray14_binary_classification import ChestXray14BinaryClassification
 from .ecgqa_preprocess import ECGQA
-from .ptbxl_diagnostic_classification import PTBXLDiagnosticClassification
+from .ptbxl_resampling import PTBXLResamplingTask
 from .chestxray14_multilabel_classification import ChestXray14MultilabelClassification
 from .covid19_cxr_classification import COVID19CXRClassification
 from .dka import DKAPredictionMIMIC4, T1DDKAPredictionMIMIC4
